admin_panel/pyproject.toml
[project]
name = "admin-panel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "pycryptodomex>=3.23.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "werkzeug>=3.1.3",
]
